style: 문자열 Constant파일에서 불러오기
